rework NxRunTargetAction to be inline with other UX & no runBlocking
